% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_using_tscores-REY123.R
\name{REY123}
\alias{REY123}
\title{REY123 Test Scores}
\usage{
REY123(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{REY123}.
}
\description{
Create a \code{REY123} object to hold the sum of REY1REC, REY2REC, and REY3REC scores.
}
