% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_scores.R
\name{scale_scores}
\alias{scale_scores}
\title{Scale Scores}
\usage{
scale_scores(scores)
}
\arguments{
\item{scores}{An object of class \code{npsych_scores}.}
}
\value{
A numeric value. The function will error if \code{scores} is not of class
\code{npsych_scores} or if its subclass is not one of \code{names(ss)}.
}
\description{
Get the scaled scores used in the T-scores.
}
