% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS.R
\docType{data}
\name{ss}
\alias{ss}
\title{Scaled Scores lookup tables from Mayo Normative Studies}
\format{
A named list of named vectors such that \code{ss$var_name[["raw_score"]]}
gives the scaled score for the variable \code{var_name} corresponding to a
raw score of \code{raw_score}.
}
\usage{
ss
}
\description{
Scaled scores from Mayo Normative Studies for a variety of variables. To get
a list of all variables, run \code{names(ss)}. Note, these names refer to variable
names in the Mayo Normative Study.
}
\section{Scaled Scores}{
Scaled scores are used in the calculation of T-scores. In short, the T-scores
are based on a regression model that fits the scaled scores as a function of
sex, age, and years of education (with some second order terms). The purpose
of the scaled scores is to force the data to be more normal looking. For more
details, see \insertRef{karstensMayoNormativeStudies2023}{ntrsTscores}.
}

\references{
\insertRef{karstensMayoNormativeStudies2023}{ntrsTscores}
}
\keyword{datasets}
