% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_using_tscores-REY123.R
\name{std_using_tscores-REY123}
\alias{std_using_tscores-REY123}
\title{REY123 T-score}
\arguments{
\item{scores}{\code{REY123} scores object}

\item{age}{numeric vector with age}

\item{sex}{numeric vector of \code{1}'s (male) and \code{2}'s (female)}

\item{educ}{numeric vector giving years of education}
}
\value{
...
}
\description{
REY123 T-score
}
