% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_using_tscores-REY5REC.R
\name{std_using_tscores-REY5REC}
\alias{std_using_tscores-REY5REC}
\title{REY5REC T-score}
\arguments{
\item{scores}{\code{REY5REC} scores object}

\item{age}{numeric vector with age}

\item{sex}{numeric vector of \code{1}'s (male) and \code{2}'s (female)}

\item{educ}{numeric vector giving years of education}
}
\value{
...
}
\description{
REY5REC T-score
}
