% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_using_tscores-REYAREC.R
\name{std_using_tscores-REYAREC}
\alias{std_using_tscores-REYAREC}
\title{REYAREC T-score}
\arguments{
\item{scores}{\code{REYAREC} scores object}
}
\value{
...
}
\description{
REYAREC T-score
}
