% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_using_tscores-REYDLIST.R
\name{std_using_tscores-REYDLIST}
\alias{std_using_tscores-REYDLIST}
\title{REYDLIST T-score}
\arguments{
\item{scores}{\code{REYDLIST} scores object}

\item{age}{numeric vector with age}

\item{sex}{numeric vector of \code{1}'s (male) and \code{2}'s (female)}

\item{educ}{numeric vector giving years of education}
}
\value{
...
}
\description{
REYDLIST T-score
}
