% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_using_tscores-REY1REC.R,
%   R/std_using_tscores-REY2REC.R, R/std_using_tscores-REYTOTAL.R
\name{std_using_tscores-REYTOTAL}
\alias{std_using_tscores-REYTOTAL}
\title{REY1REC T-score}
\arguments{
\item{scores}{\code{REYTOTAL} scores object}

\item{age}{numeric vector with age}

\item{sex}{numeric vector of \code{1}'s (male) and \code{2}'s (female)}

\item{educ}{numeric vector giving years of education}
}
\value{
...

...

...
}
\description{
REY1REC T-score

REY2REC T-score

REYTOTAL T-score
}
