% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_using_tscores.R
\name{std_using_tscores}
\alias{std_using_tscores}
\title{Standardize using T-scores}
\usage{
std_using_tscores(scores, age, sex, educ)
}
\arguments{
\item{scores}{A numeric vector of class \code{npscyh_scores} with subclass}

\item{age}{numeric vector with age}

\item{sex}{numeric vector of \code{1}'s (male) and \code{2}'s (female)}

\item{educ}{numeric vector giving years of education}
}
\value{
The standardized value of \code{scores}, whose exact type depends on the specific S7 method called
}
\description{
Standardize using T-scores
}
