% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrsTscores-package.R
\docType{package}
\name{ntrsTscores-package}
\alias{ntrsTscores}
\alias{ntrsTscores-package}
\title{ntrsTscores: Extending ntrs with T-scores}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Ralph Møller Trane \email{rmtrane@gmail.com} (\href{https://orcid.org/0000-0001-7824-1785}{ORCID})

}
\keyword{internal}
